Crate rusoto_iot_data

AWS IoT
AWS IoT-Data enables secure, bi-directional communication between Internet-connected things (such as sensors, actuators, embedded devices, or smart appliances) and the AWS cloud. It implements a broker for applications and things to publish messages over HTTP (Publish) and retrieve, update, and delete thing shadows. A thing shadow is a persistent representation of your things and their state in the AWS cloud.

If you're using the service, you're probably looking for IotDataClient and IotData.

Structs

DeleteThingShadowRequest
The input for the DeleteThingShadow operation.

DeleteThingShadowResponse
The output from the DeleteThingShadow operation.

GetThingShadowRequest
The input for the GetThingShadow operation.

GetThingShadowResponse
The output from the GetThingShadow operation.

IotDataClient
A client for the AWS IoT Data Plane API.

PublishRequest
The input for the Publish operation.

UpdateThingShadowRequest
The input for the UpdateThingShadow operation.

UpdateThingShadowResponse
The output from the UpdateThingShadow operation.

Enums

DeleteThingShadowError
Errors returned by DeleteThingShadow

GetThingShadowError
Errors returned by GetThingShadow

PublishError
Errors returned by Publish

UpdateThingShadowError
Errors returned by UpdateThingShadow

Traits

IotData
Trait representing the capabilities of the AWS IoT Data Plane API. AWS IoT Data Plane clients implement this trait.